# src/main/resources/static/openapi.yml
openapi: 3.0.1
info:
  title: Beneficiary Management API
  description: API para gerenciar beneficiários e seus documentos.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
paths:
  /beneficiaries:
    post:
      operationId: createBeneficiary
      summary: Cadastrar um beneficiário junto com seus documentos
      requestBody:
        $ref: '#/components/requestBodies/CreateBeneficiaryRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeneficiaryResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: BENEFICIARY-SERVICE-9999
                message: Error response message, related to the validation or generic errors.
                details: []
                date: 2024-06-13T18:00:01Z
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Listar todos os beneficiários cadastrados
      operationId: listAllBeneficiaries
      responses:
        '200':
          $ref: '#/components/responses/BeneficiaryList'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: BENEFICIARY-SERVICE-9999
                message: Error response message, related to the validation or generic errors.
                details: []
                date: 2024-06-13T18:00:01Z
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/{beneficiaryId}:
    get:
      operationId: getBeneficiaryDocuments
      summary: Listar todos os documentos de um beneficiário a partir de seu id
      parameters:
        - name: beneficiaryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/DocumentList'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: BENEFICIARY-SERVICE-9999
                message: Error response message, related to the validation or generic errors.
                details: []
                date: 2024-06-13T18:00:01Z
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /beneficiaries/{beneficiaryId}:
    put:
      operationId: updateBeneficiary
      summary: Atualizar os dados cadastrais de um beneficiário
      parameters:
        - name: beneficiaryId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        $ref: '#/components/requestBodies/UpdateBeneficiaryRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeneficiaryResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: BENEFICIARY-SERVICE-9999
                message: Error response message, related to the validation or generic errors.
                details: []
                date: 2024-06-13T18:00:01Z
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteBeneficiary
      summary: Remover um beneficiário
      parameters:
        - name: beneficiaryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/BeneficiaryDeleted'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: BENEFICIARY-SERVICE-9999
                message: Error response message, related to the validation or generic errors.
                details: []
                date: 2024-06-13T18:00:01Z
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  requestBodies:
    CreateBeneficiaryRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BeneficiaryRequest'
    UpdateBeneficiaryRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdateBeneficiaryRequest'

  responses:
    BeneficiaryDeleted:
      description: Beneficiário removido com sucesso
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DeleteResponse'
    BeneficiaryList:
      description: Lista de beneficiários
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/BeneficiarySummary'
    DocumentList:
      description: Lista de documentos
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Document'
  schemas:
    BeneficiaryRequest:
      type: object
      properties:
        nome:
          type: string
        telefone:
          type: string
        dataNascimento:
          type: string
          format: date
          example: "1999-03-24"
        documentos:
          type: array
          items:
            $ref: '#/components/schemas/DocumentRequest'
      required:
        - nome
        - telefone
        - dataNascimento
        - documentos
    UpdateBeneficiaryRequest:
      type: object
      properties:
        nome:
          type: string
        telefone:
          type: string
        dataNascimento:
          type: string
          format: date
          example: "1999-03-24"
      required:
        - nome
        - telefone
        - dataNascimento
    BeneficiaryResponse:
      type: object
      properties:
        id:
          type: string
        dataInclusao:
          type: string
        dataAtualizacao:
          type: string
        nome:
          type: string
        telefone:
          type: string
        dataNascimento:
          type: string
          format: date
        documentos:
          type: array
          items:
            $ref: '#/components/schemas/Document'
    BeneficiarySummary:
      type: object
      properties:
        id:
          type: string
        nome:
          type: string
    DocumentRequest:
      type: object
      properties:
        tipoDocumento:
          type: string
        descricao:
          type: string
      required:
        - tipoDocumento
        - descricao
    Document:
      type: object
      properties:
        id:
          type: string
        tipoDocumento:
          type: string
        descricao:
          type: string
      required:
        - tipoDocumento
        - descricao
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
        releaseAt:
          type: string
          format: date-time
          example: '2023-01-01T18:00:01Z'
    DeleteResponse:
      type: object
      properties:
        deletedAt:
          type: string
          format: date-time
          description: date and time the record was deleted, in UTC standard
          example: '2023-01-01T18:00:01Z'
        message:
          type: string